requirements.txt
irc~=20.4.0
tempora~=5.5.1
pip~=23.2.1
wheel~=0.41.2
pytz~=2024.1
mysql-connector-python~=8.3.0
zipp~=3.18.1
setuptools~=68.2.0
autocommand~=2.2.2
inflect~=7.2.0
typeguard~=4.2.1
mysqlclient~=2.2.4
packaging~=24.0